requirements.txt
# Core async + HTTP libraries
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
attrs==25.3.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
multidict==6.6.4
sniffio==1.3.1
yarl==1.20.1

# Google + Auth stack
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
proto-plus==1.26.1
protobuf==6.32.0
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-groq==0.3.7
langchain-text-splitters==0.3.9
langsmith==0.4.15

# Serialization / Parsing
annotated-types==0.7.0
charset-normalizer==3.4.3
dataclasses-json==0.6.7
marshmallow==3.26.1
mypy_extensions==1.1.0
orjson==3.11.2
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
PyYAML==6.0.2

# Utilities
cachetools==5.5.2
certifi==2025.8.3
distro==1.9.0
filetype==1.2.0
propcache==0.3.2
requests==2.32.5
requests-toolbelt==1.0.0
tenacity==9.1.2
urllib3==2.5.0
zstandard==0.24.0

# Data + ML stack
numpy>=1.25
pandas>=1.5
scikit-learn
lightgbm
xgboost
joblib
shap
sentence-transformers
transformers
torch
hdbscan
faiss-cpu==1.12.0
matplotlib

# PDF / Docs
pypdf==6.0.0
python-dotenv==1.1.1
beautifulsoup4

# Web / API
fastapi
uvicorn
playwright

# Logging
loguru


streamlit
pandas
matplotlib
seaborn
joblib
scikit-learn
numpy


plotly           # For interactive charts
transformers     # If using Hugging Face for sentiment analysis
wordcloud        # For visualizing frequent review terms
Pillow           # For image display or manipulation